Pure refactor to extract analyze() from PackageTypingsGenerator into a new EntryTable class
